Rename some modules.
